feature/added a common navigation layout for all users but is determined by the role and also made a cart context for the user when they add an item in the cart
